# fastapi_backend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "北京健康评估系统API",
    "description": "北京健康评估系统的后端API接口",
    "version": "1.0.0"
  },
  "paths": {
    "/api/token": {
      "post": {
        "tags": ["认证"],
        "summary": "用户登录接口，获取访问令牌",
        "operationId": "login_for_access_token",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": ["认证"],
        "summary": "用户登录接口，简化版",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误"
          }
        }
      }
    },
...
    "/api/data/": {
      "post": {
        "tags": ["数据管理"],
        "summary": "上传数据文件",
        "operationId": "create_data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "personnel_id": {
                    "type": "string"
                  },
                  "personnel_name": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["personnel_id", "personnel_name", "file"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "数据上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data"
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          }
        }
      },
      "get": {
        "tags": ["数据管理"],
        "summary": "获取数据列表",
        "operationId": "read_data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "跳过的记录数",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返回的最大记录数",
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "name": "personnel_id",
            "in": "query",
            "description": "按人员ID筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "personnel_name",
            "in": "query",
            "description": "按人员名称筛选",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取数据列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Data"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          }
        }
      }
    },
    "/api/data/{data_id}": {
      "get": {
        "tags": ["数据管理"],
        "summary": "获取特定数据",
        "operationId": "read_data_by_id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "data_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data"
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "数据不存在"
          }
        }
      },
      "delete": {
        "tags": ["数据管理"],
        "summary": "删除数据",
        "operationId": "delete_data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "data_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功删除数据"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "数据不存在"
          }
        }
      }
    },
...
    "/api/logs/": {
      "get": {
        "tags": ["日志管理"],
        "summary": "获取系统日志",
        "operationId": "read_logs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "description": "开始日期（格式：YYYY-MM-DD）",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "结束日期（格式：YYYY-MM-DD）",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "username",
            "in": "query",
            "description": "按用户名筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "按日志级别筛选（INFO, WARNING, ERROR, DEBUG）",
            "schema": {
              "type": "string",
              "enum": ["INFO", "WARNING", "ERROR", "DEBUG"]
            }
          },
          {
            "name": "skip",
            "in": "query",
            "description": "跳过的记录数",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返回的最大记录数",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取日志列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "timestamp": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "level": {
                        "type": "string"
                      },
                      "username": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Token": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["username", "password"]
      },
...
      "Data": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "personnel_id": {
            "type": "string"
          },
          "personnel_name": {
            "type": "string"
          },
          "file_path": {
            "type": "string"
          },
          "file_type": {
            "type": "string"
          },
          "upload_time": {
            "type": "string",
            "format": "date-time"
          },
          "upload_user": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 